Collect template includes when collecting resources
Template includes should also be collected when collecting resources, so
that an update in an include results in an update of the site when
auto-rebuilding.
However, there is no need to include the directories in the resources;
the directory information is already included in the path of the files,
so if a change occurs, for example, in the name of the directory, it
will be reflected in the filepath.

[#10]