Revisit the datetime definitions in the posts
Instead of redefining the entire date string in the post files and
potentially having conflicts with the filenames, it is simpler to only
allow for extra definitions of the time and timezone instead.
Additionally, the user can set a default timezone in the config file, to
avoid having to type it in on every post.

[#8]