Add support for permalinks
As described in the issue, pages should now support permalinks. There
should be enough checks in place to avoid things like generating pages
outside of the build directory, and warnings to inform user of potential
page conflicts. The documentation is also updated to inform the user on
how to use permalinks.

[#16]